+++ src/fisca/calculettes/impot-revenu/revenus-categoriels/plus-values/plus-values.test.js
@@ -10,7 +10,7 @@ describe('Fisca.Calculettes.ImpotRevenuCalculette.RevenusCategoriels.PlusValues'
         ts.dateAcquisition = new Date(2015,0,1);
         ts.dateCession = new Date(2016,0,1);
         ts.prixRevient = 100000;
-        ts.prixCession = 150000;
+        ts.revenuBrut = 150000;
 
         assert.equal(ts.plusValueBrute, 50000);
         assert.equal(ts.revenuNet, 50000);
@@ -24,12 +24,12 @@ describe('Fisca.Calculettes.ImpotRevenuCalculette.RevenusCategoriels.PlusValues'
         ts.dateAcquisition = new Date(2015,0,1);
         ts.dateCession = new Date(2025,0,1);
         ts.prixRevient = 100000;
-        ts.prixCession = 500000;
+        ts.revenuBrut = 500000;
 
         assert.equal(ts.plusValueBrute, 400000);
         assert.equal(ts.revenuNet, 0);
 
-        ts.prixCession = 1500000;
+        ts.revenuBrut = 1500000;
         assert.equal(ts.plusValueBrute, 1500000 - 100000);
         assert.equal(ts.revenuNet, Math.round((1 - 0.85) * (900000)));
         
@@ -42,7 +42,7 @@ describe('Fisca.Calculettes.ImpotRevenuCalculette.RevenusCategoriels.PlusValues'
         ts.dateAcquisition = new Date(2015,0,1);
         ts.dateCession = new Date(2025,0,1);
         ts.prixRevient = 100000;
-        ts.prixCession = 500000;
+        ts.revenuBrut = 500000;
 
         assert.equal(ts.plusValueBrute, 400000);
         assert.equal(ts.revenuNet, Math.round(400000 * (1 - 0.85)));